% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro_data_processing.R
\name{ceh.landcov.extract}
\alias{ceh.landcov.extract}
\title{Function for extracting CEH landcover classess to grid cells}
\usage{
ceh.landcov.extract(
  ceh_dat,
  region_sf,
  hexCellArea,
  hexGrid = NULL,
  buffer = 1
)
}
\arguments{
\item{ceh_dat}{A CEH Land cover dataset (e.g. 2019) as an sf object.}

\item{region_sf}{An sf object defining the spatial region.}

\item{hexCellArea}{A numeric value specifying the area of the hexagonal cell (in km2).}

\item{hexGrid}{A pre-made hexagonal grid, as produced by \code{makeHexGrid}.}

\item{buffer}{The buffer distance in kilometers.}
}
\value{
A data frame with the area of cover from each land class category (LC_1 to LC_21).
}
\description{
LC_1 = Deciduous woodland (wd_dec)
LC_2 = Coniferous woodland (wd_con)
LC_3 = Arable (arab)
LC_4 = Improved grassland (grImp)
LC_5 = Neutral grassland (grNeu)
LC_6 = Calcareous grassland (grCal)
LC_7 = Acid grassland (grAci)
LC_8 = Fen (fen)
LC_9 = Heather (heath)
LC_10 = Heather grassland (grHeath)
LC_11 = Bog (bog)
LC_12 = Inland rock (rkInld)
LC_13 = Saltwater (saltwat)
LC_14 = Freshwater (freshwat)
LC_15 = Supralittoral rock (rkSup)
LC_16 = Supralittoral sediment (sediSup)
LC_17 = Littoral rock (rkLit)
LC_18 = Littoral sediment (sediLit)
LC_19 = Saltmarsh (saltmar)
LC_20 = Urban (urban)
LC_21 = Suburban (surban)
}
