% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro_data_processing.R
\name{climate.hex.variation}
\alias{climate.hex.variation}
\title{Calculate the mean and standard deviation of bioclimate variables across each cell}
\usage{
climate.hex.variation(
  clim_grd,
  region_sf,
  hexCellArea,
  hexGrid = NULL,
  var_name,
  buffer_dist
)
}
\arguments{
\item{clim_grd}{A grid of 100 m x 100 m micro-bioclimate variables}

\item{region_sf}{An sf object defining the spatial region.}

\item{hexCellArea}{A numeric value specifying the area of the hexagonal cell (in km2).}

\item{hexGrid}{A pre-made hexagonal grid, as produced by \code{makeHexGrid}.}

\item{var_name}{Name of climate variable.}

\item{buffer_dist}{Extract values from buffer if presence in focal cell is likely to be
affected by availability of surrounding climate}
}
\value{
A tibble data frame.
}
\description{
Calculate the mean and standard deviation of bioclimate variables across each cell
}
