% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_processing.R
\name{db.transform.proj}
\alias{db.transform.proj}
\title{Re-project point data}
\usage{
db.transform.proj(bioDat, crs_sf_ref, region = NULL)
}
\arguments{
\item{bioDat}{A data frame with SpatialLong given as \code{X} and SpatialLat given as \code{Y}.}

\item{crs_sf_ref}{An sf object with the desired crs.}

\item{region}{A region defined within  \code{crs_sf_ref}.}
}
\value{
An sf object.
}
\description{
This function uses a reference spatial polygon to transform the point data to a new projection
and returns an sf spatial object.
}
