% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro_data_processing.R
\name{distance.decay}
\alias{distance.decay}
\title{Process spatial data from OS and other sources}
\usage{
distance.decay(enviro_sf, region_sf, hexCellArea, hexGrid = NULL, var_name)
}
\arguments{
\item{enviro_sf}{An environmental sf object}

\item{region_sf}{An sf object defining the spatial region.}

\item{hexCellArea}{A numeric value specifying the area of the hexagonal cell (in km2).}

\item{hexGrid}{A pre-made hexagonal grid, as producted by \code{makeHexGrid}.}

\item{var_name}{Name of environmental variable from \code{enviro_sf}.}
}
\value{
An sf object.
}
\description{
Distance decay to environmental features
}
