% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro_data_processing.R
\name{extract.bioclim}
\alias{extract.bioclim}
\title{Bioclimate data is from 1983 to 2017 - need to aggregate}
\usage{
extract.bioclim(dir_name)
}
\arguments{
\item{dir_name}{The path to the climate files}
}
\value{
A list of data frames
}
\description{
add a period argument to select years
}
