% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enviro_data_processing.R
\name{feature.occurence}
\alias{feature.occurence}
\title{Count feature occurrence within cells}
\usage{
feature.occurence(enviro_sf, region_sf, hexCellArea, hexGrid = NULL, var_name)
}
\arguments{
\item{enviro_sf}{An environmental sf object}

\item{region_sf}{An sf object defining the spatial region.}

\item{hexCellArea}{A numeric value specifying the area of the hexagonal cell (in km2).}

\item{hexGrid}{A pre-made hexagonal grid, as producted by \code{makeHexGrid}.}

\item{var_name}{Name of environmental variable from \code{enviro_sf}.}
}
\value{
An sf object.
}
\description{
Count feature occurrence within cells
}
