% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_processing.R
\name{focal.species.select}
\alias{focal.species.select}
\title{Focal species select}
\usage{
focal.species.select(
  bioDat,
  focal_sp,
  region_sf,
  grain,
  hexGrid = NULL,
  collapse_time = TRUE
)
}
\arguments{
\item{bioDat}{A data frame with SpatialLong given as X and SpatialLat given as Y.}

\item{focal_sp}{A focal species to select.}

\item{region_sf}{An sf object defining the spatial region.}

\item{grain}{Area of each hexagonal grid cell.}

\item{hexGrid}{A pre-made hexagonal grid, as produced by \code{makeHexGrid}.}

\item{collapse_time}{Should all year's data be collapsed into a single presences/absences data frame}
}
\value{
A tibble data frame holding data for \code{focal_sp}.
}
\description{
Focal species select
}
