% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_processing.R
\name{makeHexGrid}
\alias{makeHexGrid}
\title{Make hexagonal grid}
\usage{
makeHexGrid(region_sf, hexCellArea)
}
\arguments{
\item{region_sf}{An sf object defining the spatial region.}

\item{hexCellArea}{A numeric value specifying the area of the hexagonal cell (in km2).}
}
\value{
An sf object with 'id' used as a unique cell id.
}
\description{
Analysis is conducted on hexagonal grids and this function creates the grid with a specified area.
}
