% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_model.R
\name{occupancy.multi}
\alias{occupancy.multi}
\title{Create multi-species occupancy distribution data}
\usage{
occupancy.multi(
  enviro_dat,
  enviro_vars,
  n_spp = 50,
  pca_axes = 3,
  region_sf,
  plotRichness = FALSE,
  plotSpecies = FALSE
)
}
\arguments{
\item{enviro_dat}{Environmental data (requires \code{id}, \code{X}, \code{Y},
plus some environmental variables).}

\item{enviro_vars}{A named vector of environmental variables to be included.}

\item{n_spp}{The number of species to simulate.}

\item{pca_axes}{The number of pca axes to use in creating a species.}

\item{region_sf}{An sf object defining the spatial region.}

\item{plotRichness}{Create richness plot.}

\item{plotSpecies}{Create example species plot (plots species number 1).}
}
\value{
A data frame
}
\description{
The function create a data frame of species occupancy for a defined number of
species by defining the probability of occurrence as a function of Gaussian
responses to the axes of the principle components of the environmental data.
}
