% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package.load}
\alias{package.load}
\title{Load and, if necessary, install packages automatically}
\usage{
package.load(packageNames)
}
\arguments{
\item{packageNames}{A vector of quoted package names.}
}
\value{
A table of packages installed and loaded.
}
\description{
Just saves a bunch of time, especially when moving between
computers.
}
