% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_model.R
\name{plot.virtual.species}
\alias{plot.virtual.species}
\title{Plot species and sampling}
\usage{
plot.virtual.species(multi_occ, multi_det, sp = 1, year = 1)
}
\arguments{
\item{multi_occ}{The output from \code{occupancy.multi}.}

\item{multi_det}{The output from \code{virual.sampling}.}

\item{sp}{A numeric value indicating the species to plot.}

\item{year}{A numeric value indicating the year to plot.}
}
\value{
A map of the 'real' distribution and the sample, with sampling
weighting as shown as the background.
}
\description{
Plot species and sampling
}
