% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_processing.R
\name{recToHexSetUp}
\alias{recToHexSetUp}
\title{Assign to hexagonal grid by any time slice}
\usage{
recToHexSetUp(bioDat, timeSlice, region_sf, hexCellArea, hexGrid = NULL)
}
\arguments{
\item{bioDat}{A data frame with SpatialLong given as \code{X} and SpatialLat given as \code{Y}.}

\item{timeSlice}{A numeric value specifying the length on each timeslice (in years).}

\item{region_sf}{An sf object defining the spatial region.}

\item{hexCellArea}{A numeric value specifying the area of the hexagonal cell (in km2).}

\item{hexGrid}{A pre-made hexagonal grid, as producted by \code{makeHexGrid}.}
}
\value{
A tibble data frame with hexagonal cell id and X, Y coordinates.
}
\description{
Function assigns records to grid cells based on spatial overlap.
}
