% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_processing.R
\name{recXgridXtime}
\alias{recXgridXtime}
\title{Number of records \code{x} hexagonal grid \code{x} time slice}
\usage{
recXgridXtime(
  bioDat,
  sppGrp,
  timeSlice,
  region_sf,
  hexCellArea,
  bioHex = NULL,
  hexGrid = NULL
)
}
\arguments{
\item{bioDat}{A data frame with SpatialLong given as X and SpatialLat given as Y.}

\item{sppGrp}{Name of species group (unwise to run on all species!!!).}

\item{timeSlice}{A numeric value specifying the length on each timeslice (in years).}

\item{region_sf}{An sf object defining the spatial region.}

\item{hexCellArea}{A numeric value specifying the area of the hexagonal cell (in km2).}

\item{bioHex}{An export from \code{recToHexSetUp}.}

\item{hexGrid}{A pre-made hexagonal grid, as producted by \code{makeHexGrid}.}
}
\value{
A data frame with hexagonal cell id and X, Y coordinates.
}
\description{
This function summarises number of records per cell and time slice.
This is useful for exploring trends in record collection for identifying periods when
records may be more or less representative.
}
