% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_processing.R
\name{siteXSpeciesXTime}
\alias{siteXSpeciesXTime}
\title{Site \code{x} Species \code{x} Time data frame}
\usage{
siteXSpeciesXTime(
  bioDat,
  timeSlice,
  region_sf,
  hexCellArea,
  bioHex = NULL,
  hexGrid = NULL,
  randTS = FALSE
)
}
\arguments{
\item{bioDat}{A data frame with SpatialLong given as X and SpatialLat given as Y.}

\item{timeSlice}{A numeric value specifying the length on each timeslice (in years).}

\item{region_sf}{An sf object defining the spatial region.}

\item{hexCellArea}{Area of each hexagonal grid cell.}

\item{bioHex}{An export from \code{recToHexSetUp}.}

\item{hexGrid}{A pre-made hexagonal grid, as produced by \code{makeHexGrid}.}

\item{randTS}{If TRUE the date of surveys are randomised within each cell.}
}
\value{
A binary (0,1) presence absence data frame is returned
}
\description{
This function creates a site \code{x} species \code{x} time matrix
(presence/absence), which can then be subset into specific time slices.
}
