% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source.keep}
\alias{source.keep}
\title{Source only certain functions}
\usage{
source.keep(keepFunctions, functionPath)
}
\arguments{
\item{keepFunctions}{Functions wanted in memory}

\item{functionPath}{Path to where functions are stored}
}
\description{
This helps keep on those functions needed for a particular
section of a long analysis load.
}
