% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_processing.R
\name{species.evaluation}
\alias{species.evaluation}
\title{Plots for evaluating species distributions}
\usage{
species.evaluation(
  spp_dat,
  focal_sp,
  start_yr = 2001,
  end_yr = NULL,
  region_sf,
  plotType = c("Map", "Hist", "ListLength")
)
}
\arguments{
\item{spp_dat}{An export from \code{db.transform.proj}}

\item{focal_sp}{The SpeciesVenacular name.}

\item{start_yr}{A numeric year defining the start of the focal period.}

\item{end_yr}{A numeric year defining the end of the focal period.}

\item{region_sf}{An sf object defining the spatial region.}

\item{plotType}{One of Map, Hist, or listLength.}
}
\value{
A ggplot figure.
}
\description{
Plots for evaluating species distributions
}
