% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_processing.R
\name{species.group.select}
\alias{species.group.select}
\title{Obtain species group data}
\usage{
species.group.select(
  bioDat,
  sppGrp,
  region_sf,
  regionSelect = NULL,
  sppSelect = NULL
)
}
\arguments{
\item{bioDat}{A data frame with SpatialLong given as \code{X} and SpatialLat given as \code{Y}.}

\item{sppGrp}{Name of species group (unwise to run on all species!!!).}

\item{region_sf}{An sf object defining the spatial region.}

\item{regionSelect}{A region defined within  \code{crs_sf_ref}.}

\item{sppSelect}{A SpeciesVenacular name or vector of names if just a subset of species is wanted.}
}
\value{
A filtered and transformed data frame.
}
\description{
Wrapper function for filtering by species groups and projecting point data
}
