% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_data_processing.R
\name{target.group.weights}
\alias{target.group.weights}
\title{Target group weightings for modelling}
\usage{
target.group.weights(
  sppBioDat,
  sppGrp,
  region_sf,
  grain,
  hexGrid = NULL,
  showPlot = TRUE
)
}
\arguments{
\item{sppBioDat}{A data frame with SpatialLong given as X and SpatialLat given as Y, filtered for target group species.}

\item{sppGrp}{Name of species group (Just for plot labels.}

\item{region_sf}{An sf object defining the spatial region.}

\item{grain}{Area of each hexagonal grid cell.}

\item{hexGrid}{A pre-made hexagonal grid, as produced by \code{makeHexGrid}.}

\item{showPlot}{Produce a plot.}
}
\value{
A tibble data frame holding data with target group weights.
}
\description{
Create target group for evaluating bias in data collection
and for sampling background data in presence-only modelling
}
