% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_model.R
\name{virtual.sampling}
\alias{virtual.sampling}
\title{Virtual sampling function}
\usage{
virtual.sampling(
  multi_occ,
  p_sites = 0.1,
  n_yrs = 10,
  spat_bias = TRUE,
  samp_weights = NULL,
  bias_level = 3,
  rep_visit = 2.5,
  maxVisit = 10,
  reli_score = c("low", "medium", "high")
)
}
\arguments{
\item{multi_occ}{The output from \code{occupancy.multi}.}

\item{p_sites}{The proportion of sites available for sampling.}

\item{n_yrs}{The number of years to simulate.}

\item{spat_bias}{The sampling spatially biased on random.}

\item{samp_weights}{A pre-generated data.frame of sample weights.}

\item{bias_level}{A number specifying the degree of spatial biased. The
higher the number the more spatially biased the sampling scheme.}

\item{rep_visit}{A number (exponent) controlling the distribution of repeat
visits to sites.}

\item{maxVisit}{The maximum number of visits to a site in a year.}

\item{reli_score}{How likely is an observer to record all species detected on
a visit.}
}
\value{
A data.frame of visit x species detections.
}
\description{
Virtual citizen scientists visit sites:
\itemize{
\item Ideally, visits are spatially and temporally random, they record everything
they detect, record the area surveyed.
\item and time spend in that area, and some index of their skill-level.
}
}
